+++ NEWS.md
@@ -8,6 +8,7 @@ New Features
 Bug Fixes
 
 * Previously, the `sim_n_datasets()` function used `stats::runif(1)` as a default for the `seed` argument. Because seeds are coerced to integers in `set.seed()`, this essentially meant the `seed` argument was always set to 0 (unless changed by the user), which was not intended. We changed the default to be `NULL`, which is equivalent to not setting a `seed`. This might change results obtained using previous versions. To get the same result as in previous versions, use `seed=0` or `seed=stats::runif(1)`.
+* Fixed a bug that occurred when calling `node()` or `node_td()` inside a function with objects passed to `parents` or `formula`. 
 
 Documentation